move source code to allign with latest kubebuilder
