+++ .circleci/config.yml
@@ -19,7 +19,7 @@ jobs:
       image: ubuntu-2204:2023.02.1
     environment: &versions
       # https://github.com/kubernetes/kubernetes/releases
-      KUBECTL_VERSION: v1.27.1
+      KUBECTL_VERSION: v1.26.4
       # https://hub.docker.com/r/rancher/k3s/tags
       K3S_VERSION: v1.26.1-k3s1
       # https://github.com/rancher/k3d/releases
