+++ chart/helm3/sops-secrets-operator/Chart.yaml
@@ -1,7 +1,7 @@
 apiVersion: v2
 # UPDATE_HERE
-version: 0.15.0
-appVersion: 0.9.0
+version: 0.14.2
+appVersion: 0.8.2
 type: application
 description: Helm chart deploys sops-secrets-operator
 name: sops-secrets-operator
