+++ chart/helm3/sops-secrets-operator/tests/operator_test.yaml
@@ -31,8 +31,8 @@ tests:
             app.kubernetes.io/instance: sops
             app.kubernetes.io/managed-by: Helm
             app.kubernetes.io/name: sops-secrets-operator
-            app.kubernetes.io/version: 0.9.0
-            helm.sh/chart: sops-secrets-operator-0.15.0
+            app.kubernetes.io/version: 0.8.2
+            helm.sh/chart: sops-secrets-operator-0.14.2
 
   # custom name
   - it: should correctly render custome name
@@ -170,7 +170,7 @@ tests:
       # UPDATE_HERE
       - equal:
           path: spec.template.spec.containers[0].image
-          value: isindir/sops-secrets-operator:0.9.0
+          value: isindir/sops-secrets-operator:0.8.2
       - equal:
           path: spec.template.spec.containers[0].imagePullPolicy
           value: Always
