+++ chart/helm3/sops-secrets-operator/values.yaml
@@ -13,7 +13,7 @@ image:
   # -- Operator image name
   repository: isindir/sops-secrets-operator
   # -- Operator image tag
-  tag: 0.9.0
+  tag: 0.8.2
   # -- Operator image pull policy
   pullPolicy: Always
